// src/etl/GestionCursos/Courses/courses.types.ts
export type LXPCourse = {
  id?: string;
  topic_id?: string | null;
  image_url: string;
  name: string;
  type: string;
  client_id: string;
  course_fb: string;
  stage: number;
  created_at: string;
  min_score: number;
  min_progress: number;
  competencies_json?: string | null;
  ous_json?: string | null;
  created_by_json?: string | null;
  instructors_json?: string | null;
  knowledge_json?: string | null;
  requirements_json?: string | null;
  roles_json?: string | null;
  skills_json?: string | null;
  tags_json?: string | null;
  description?: string | null;
  difficulty?: number | null;
  duration: number;
  hide: boolean;
  language: string;
  updated_at?: string | null;
  privacy?: string | null;
  welcome_message?: string | null;
  reason?: string | null;
  video_json?: string | null;
  is_deleted: boolean;
  deleted_at?: string | null;
  dc3Available?: boolean | null;
  external_course_id?: number | null;
  min_attendance?: number | null;
  block_after_due_date?: boolean;
  price?: number;
  currency?: string;
  validity?: string;
  dc4Available?: boolean;
  dc3_data_json?: string | null;
  dc4_data_json?: string | null;
  dynamic_end_days?: boolean;
  attributes_json?: string | null;
  competencies_levels?: string | null;
  show_evation_feedback?: string | null;
  instructor_signature_dc3?: string | null;
  external_url?: string | null;
  lined_courses?: string | null;
  is_lined?: boolean;
  limit_quota_ou?: number | null;
  origin?: string | null;
  completed_on_finish_attempts?: boolean;
};

export type VdmLMSCourse = {
  id?: string;
  client_id: string | null;
  author_id: string | null;
  name: string;
  description: string;
  image_url: string;
  settings: string;
  features: string;
  // related: string; // TODO: Borrar en producción
  tags: string;
  is_draft: boolean;
  is_active: boolean;
  is_private: boolean;
  stage: number;
  duration?: number | null;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
  legacy: string;
  legacy_course_fb: string;
  show_evaluation_feedback: number | null;
  language: string;
  difficulty: string;
  validity: string | null;
};

export type VdmCourse = {};